+++ tests/apps/courses/test_templates_organization_detail.py
@@ -9,7 +9,7 @@ from django.test.utils import override_settings
 from cms.api import add_plugin
 from cms.test_utils.testcases import CMSTestCase
 
-from richie.apps.core.factories import UserFactory
+from richie.apps.core.factories import PageFactory, UserFactory
 from richie.apps.courses.cms_plugins import CategoryPlugin, OrganizationPlugin
 from richie.apps.courses.factories import (
     CategoryFactory,
@@ -323,3 +323,37 @@ class OrganizationCMSTestCase(CMSTestCase):
             name=person.extended_object.get_title(),
         )
         self.assertIsNotNone(re.search(pattern, str(response.content)))
+
+    def test_template_organization_detail_without_organization(self):
+        """
+        A organization template page without attached organization should show an error banner
+        explaining to the user that he/she is misusing the template.
+        """
+        page = PageFactory(
+            template="courses/cms/organization_detail.html",
+            title__language="en",
+            should_publish=True,
+        )
+
+        with self.assertTemplateUsed(
+            "courses/cms/fragment_error_detail_template_banner.html"
+        ):
+            response = self.client.get(page.get_absolute_url())
+
+        self.assertEqual(response.status_code, 200)
+        self.assertContains(
+            response,
+            (
+                '<div class="banner banner--error banner--rounded" role="alert">'
+                '<svg class="banner__icon"><use href="#icon-cross" /></svg>'
+                '<p class="banner__message">'
+                "A organization object is missing on this organization page. "
+                "Please select another page template."
+                "<br />"
+                "If what you need is a organization page, you need to create it "
+                'via the wizard and choose "New organization page".'
+                "</p>"
+                "</div>"
+            ),
+            html=True,
+        )